refacto to card controller
