+++ backend/server.js
@@ -3,68 +3,10 @@ const cors = require('cors');
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
 const path = require('path');
-const PDFDocument = require('pdfkit');
-const SVGtoPDF = require('svg-to-pdfkit');
 const auth = require('./controllers/Auth');
-
-PDFDocument.prototype.addSVG = function addSVG(svg, x, y, options) {
-  return SVGtoPDF(this, svg, x, y, options);
-};
-
-function generateCell(x, y, content, cell) {
-  return `
-    <g key="${cell}">
-      <rect
-        x="${x}"
-        y="${y}"
-        width="194"
-        height="194"
-        fill="#fff"
-        stroke="#000"
-        strokeWidth="3"
-      />
-      <text
-        x="${x + 96}"
-        y="${y + 124}"
-        font-size="64"
-        text-anchor="middle"
-        alignment-baseline="central"
-      >
-        ${content}
-      </text>
-    </g>`;
-}
-function generateCard(w, h) {
-  const freeSpace = Math.round((w * h) / 2 + h - 1);
-  h += h;
-  const cells = [];
-  const totalCells = w * h;
-  let data;
-  let x = 0;
-  let y = 0;
-  let bingoStr = 'BINGO';
-  let cell = 0;
-  for (let i = 0; i < h; i++) {
-    for (let j = 0; j < w; j++) {
-      if (cell < 5) {
-        data = bingoStr[cell];
-      } else {
-        data = cell === freeSpace ? 'FREE' : Math.floor(Math.random() * 100);
-      }
-      if (data < 10) {
-        data = '0' + data;
-      }
-      x = ((j + 1) * 200);
-      y = 200 * (i + 1);
-      cells.push(generateCell(x, y, data, cell));
-      cell++;
-    }
-  }
-  return `<svg id="preview" viewBox="0 0 1400 1400">${cells.join('')}</svg>`;
-}
+const card = require('./controllers/Card');
 
 const DB_URL = process.env.MONGODB_URI || 'mongodb://localhost:27017/bingo';
-
 const server = express();
 
 server.use(cors({
@@ -84,33 +26,11 @@ server.use(express.static(path.join(__dirname, '../client/build')));
 server.post('/auth/register', auth.register);
 server.post('/auth/login', auth.login);
 server.post('/auth/reset', auth.reset);
-
-server.get('/cards', (req, res) => {
-  res.json([]);
-});
-
-server.get('/card/:id', (req, res) => {
-  res.json({});
-});
-
-server.post('/card/create', (req, res) => {
-  res.json({});
-});
-
-server.post('/card/edit', (req, res) => {
-  res.json({});
-});
-
-server.get('/cards/download', (req, res) => {
-  const doc = new PDFDocument();
-  doc.pipe(res);
-  for(let i = 0; i < Math.floor(Math.random() * 20); i++) {
-    SVGtoPDF(doc, generateCard(5, 5), 0, 0);
-    doc.addPage();
-  }
-  doc.end();
-});
-
+server.get('/cards', card.getAll);
+server.get('/card/:id', card.get);
+server.post('/card/create', card.create);
+server.post('/card/edit', card.edit);
+server.get('/cards/download', card.download);
 server.get('*', (req, res) => {
   res.sendFile(path.join(__dirname, '../client/build/index.html'));
 });
